wahoo: disable apex updating
APEXs have quite a bit of overhead in updatable form: they are ext4
filesystems mounted using loopback block devices, stored on yet another
ext4 filesystem (system or vendor). They also increase boot time
substantially due to needing to set up and mount 27 images with AVB. On
a Pixel 2 XL, apexd-bootstrap took 485 ms during boot, and apexd took
471 ms for a total of 955 ms — nearly a second — spent on APEX setup,
which is quite significant for modern devices. At least 422 ms out of
that was spent blocking the rest of the system's initialization, but the
overall slowdown is likely greater due to increased system load.

Updatable APEXs likely also come with a non-negligible amount of memory
overhead due to setting up and mounting so many filesystem images, as
well as AVB and potential double-caching effects caused by having two
layers of filesystems in use.

Being able to update APEXs is useless for us anyway because we don't
use Google's prebuilt modules and we have no other way to deliver APEX
updates. Even if we were able to use Google's modules, we still want to
be able to modify many of the APEX modules, so we need to build them
from source.